Moved "types/graphql.ts" to "graphql/types.ts" and also "types/context.ts" to "misc/context-type.ts", so that I could delete the folder "types" (since I was declaring types in other places as well)
