+++ codegen.yml
@@ -2,7 +2,7 @@ overwrite: true
 schema: 'http://localhost:4000/graphql'
 documents: null
 generates:
-  src/types/graphql.ts:
+  src/graphql/types.ts:
     plugins:
       - 'typescript'
       - 'typescript-resolvers'
