+++ src/graphql/main/main-resolvers.ts
@@ -1,5 +1,5 @@
-import { Context } from '../../types/context';
-import { Resolvers } from '../../types/graphql';
+import { Context } from '../../misc/context';
+import { Resolvers } from '../types';
 
 // Resolvers for common fields and types declared in Query and Mutation
 export const mainResolvers: Resolvers<Context> = {
