+++ src/misc/context-type.ts
@@ -1,7 +1,7 @@
 import { UserModel } from '../db-models/user-model';
 import { UserModelConverter } from '../model-converters/user-model-converter';
 import { IRepo } from '../repositories/irepo';
-import { UserCreationInput, UserUpdateInput } from './graphql';
+import { UserCreationInput, UserUpdateInput } from '../graphql/types';
 
 export type DataSources = {
 	userRepo: IRepo<UserModel, UserCreationInput, UserUpdateInput>;
